changes to read from FV-1 eeprom

+++ comms/i2c/i2c.h
@@ -29,6 +29,6 @@ extern result_t i2c_restart(enum i2c_channel chan, void (*callback)(result_t));
 extern result_t i2c_stop(enum i2c_channel chan);
 
 extern result_t i2c_write(enum i2c_channel chan, uint8_t *tx_buf, uint8_t num_tx_bytes, void (*callback)(result_t));
-extern result_t i2c_read(enum i2c_channel chan, uint8_t *rx_buf, uint8_t num_rx_bytes, void (*callback)(result_t));
+extern result_t i2c_read(enum i2c_channel chan, uint8_t *rx_buf, uint16_t num_rx_bytes, void (*callback)(result_t));
 
 #endif // SYS_I2C
