+++ comms/i2c/mc24LC64.h
@@ -2,6 +2,6 @@
 
 #ifdef SYS_24LC64
 
-extern result_t mc24lc64_read(enum i2c_channel, uint8_t chip_addr, uint16_t mem_address, uint8_t num_bytes, uint8_t *buffer, void (*callback)(result_t, uint8_t *, uint8_t));
+extern result_t mc24lc64_read(enum i2c_channel, uint8_t chip_addr, uint16_t mem_address, uint16_t num_bytes, uint8_t *buffer, void (*callback)(result_t, uint8_t *, uint8_t));
 
 #endif // SYS_24LC64
